c2.38-feedback_jschmidt
c2.38-40 - if a GOAL is an ideal future reality with a PLAN to increase its likelihood (38), then everything that could be true in your life and isn’t (40) ISNT a goal is it? (confirmed in 41)
     (43) isn’t having a PLAN (necessitated in 38) part of it being a GOAL?
answer: clarified 2.38 that this is a class definition about what is not true, not what you'd like your goal to be.